+++ src/components/features/modal/ModalBase.tsx
@@ -1,4 +1,3 @@
-"use client";
 
 import {
   Dialog,
